+++ client/src/app/(auth)/products/insights/project/page.tsx
@@ -59,7 +59,7 @@ export default function Page() {
             <Hero>
               <p className="branding">{project?.data?.project?.name}</p>
               <p className="muted-text mt-3">Your Project Analytics will be displayed below (if any)</p>
-              <SensitiveInfoPanel credentialIcon={<CubeIcon />} credentialName="Project ID" credentialValue={project?.data?.project?.projectId} />
+              <SensitiveInfoPanel credentialIcon={<CubeIcon />} credentialName="Project ID" credentialValue={project?.data?.project?._id} />
               <SensitiveInfoPanel credentialIcon={<LockOpen2Icon />} credentialName="Project Passkey" credentialValue={project?.data?.project?.projectPasskey} />
               <Button variant="danger" onClick={deleteProject}>Delete Project<TrashIcon className="icon-right" /></Button>
             </Hero>