+++ src/api/products/insights/insights.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, BadRequestException, NotFoundException } from "@nestjs/common"
-import { randomBytes } from "crypto"
+import { randomUUID } from "crypto"
 import { CreateAnalyticsDto } from "./dto/create-analytics.dto"
 import { CreateProjectDto } from "./dto/create-project.dto"
 import { InsightsRepository } from "./insights.repository"
@@ -14,9 +14,8 @@ export class InsightsService {
       const count = await this.insightsRepository.countProjects(workspaceId)
 
       if (count < 10) {
-        const projectId = randomBytes(16).toString("hex")
-        const projectPasskey = randomBytes(32).toString("hex")
-        const project = await this.insightsRepository.createProject(workspaceId, name, projectId, projectPasskey)
+        const projectPasskey = randomUUID()
+        const project = await this.insightsRepository.createProject(workspaceId, name, projectPasskey)
         return project
       }
 
@@ -67,7 +66,7 @@ export class InsightsService {
       const { workspaceId: workspaceIdFromProject } = project
       if (workspaceIdFromProject.toString() === workspaceId) {
         await this.insightsRepository.deleteProjectById(workspaceId, projectId)
-        return true
+        return { success: true }
       }
 
       else {
@@ -87,7 +86,7 @@ export class InsightsService {
       if (project.workspaceId.toString() === workspaceId) {
         const projectId = project.id
         await this.insightsRepository.createAnalytics(workspaceId, projectId, component, event, info, statusCode)
-        return true
+        return { success: true }
       }
 
       else {