+++ src/api/subscription/dto/subscribe.dto.ts
@@ -1,8 +1,9 @@
 import { IsNotEmpty } from "class-validator"
+import { SubscriptionPlans } from "../subscription.config"
 
 export class SubscribeDto {
   @IsNotEmpty()
-  readonly selectedPlan: string
+  readonly selectedPlan: SubscriptionPlans
 
   @IsNotEmpty()
   readonly transactionHash: string