+++ src/constants/other-constants.ts
@@ -1,6 +1,6 @@
-import { envConfig } from "src/config/env.config"
+import { envConfig } from "src/env.config"
 
-export const otherConstants: Record<string, string> = {
+export const otherConstants = {
   tokenIssuer: `https://${envConfig.brandName}.vercel.app`,
   polygonScanApiEndpoint: "https://api-testnet.polygonscan.com"
 }